src/color_seasons/views.py
from rest_framework import viewsets

from .models import Color, Season
from .serializers import ColorSerializer, SeasonSerializer


class ColorViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Color.objects.all()
    serializer_class = ColorSerializer


class SeasonViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Season.objects.all()
    serializer_class = SeasonSerializer
