requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
alabaster==0.7.12         # via sphinx
asgiref==3.2.3            # via django
attrs==19.3.0             # via pytest
babel==2.8.0              # via sphinx
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
dj-database-url==0.5.0
django==3.0.2
djangorestframework==3.11.0
docutils==0.15.2          # via sphinx
idna==2.8                 # via requests
imagesize==1.2.0          # via sphinx
jinja2==2.10.3            # via sphinx
markupsafe==1.1.1         # via jinja2
more-itertools==8.0.2     # via pytest
packaging==20.0           # via pytest, sphinx
pip-tools==4.3.0
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.4
py==1.8.1                 # via pytest
pygments==2.5.2           # via sphinx
pyparsing==2.4.6          # via packaging
pytest==5.3.2
pytz==2019.3              # via babel, django
requests==2.22.0          # via sphinx
six==1.13.0               # via packaging, pip-tools
snowballstemmer==2.0.0    # via sphinx
sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
sqlparse==0.3.0           # via django
urllib3==1.25.7           # via requests
wcwidth==0.1.8            # via pytest
whitenoise==5.0.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools
